requirements.txt
pywebview==6.0
bottle==0.13.4
websockets==15.0.1